Catch PHP7 Throwable
Catching Throwable helps to catch a larger range of errors since PHP 7
(and avoid a worker to break while it's running).

Here I duplicated the catch block to have both Exception and Throwable
catching in order to keep the compatibility with PHP 5. The logic in the
block was then moved to a separate private method.