Allow building more exotic retry routing keys for advanced usage
A new `retry_key_generator` option allows passing a callable to build
the retry routing key in case replacing the attempt number in a static
pattern is not enough.
The callable receives the attempt number as first argument and the
processed message as second argument, and must return a string for the
retry routing key.
The default value of that option is a callable using the
`retry_key_pattern` option, which is now required only when the callable
is not provided explicitly.