Document possible redeliveries
We have been observing lots of redeliveries in our production
environment, and the only way I found to reproduce them was to use
`sleep($max_execution_time + 1)` at the beginning of my business
processor.
Increasing the max execution time resulted in a greater gap between
redeliveries, which confirms the hypothesis that this processor is
responsible for that.