Check time after running the child processor
There are two checks in the processor: on in process(), one in sleep().
In the event there are many messages in the queue when the time limit is
reached, it will be the check in process() that will kick in. Most users
will put the AckProcessor inside this one because that is what is
recommended in the docs, and will there will be no ack or nack in that
case, because of that check, which means there will be redeliveries.
By performing the check after running the child processor, we avoid the
redeliveries.