Mark all classes as becoming final in the next major version
As long as classes are marked as `@final since ...`, they are still
subject to the normal BC rules for non-final classes. The next major
version will either make them final or change the tag to `@final` which
will make them subject to the BC rules of final classes (i.e. not
considering it a change a BC break if it can only break child classes).

Classes in the library are not meant to be extended through inheritance.
Processors are naturally based on composition which is the way to
customize the library.