Drop support for unmaintained Doctrine versions
These dependencies are in the require-dev section not because they are
used for development, but because they are optional dependencies. This
means we already support recent versions of Doctrine, and it shows in
the codebase.
Let's simplify and test against maintained versions only, it will make
it easier to upgrade other packages.